remove uninstallation
Uninstallation of an automated installation with state definition
is not reasonable. There is no record of what has been changed
when a role has been applied. Therefore an uninstallation cannot be
done in the sense of just removing everything which has been installed.
Removing just everything is dangerous in the sense that it could
delete and remove resources which are still needed.