improve docker compose startup file handling
Systemd service is directly started from the initial docker-compose
file and the permanent one copied over the initial one afterwards.
This adds the advantage that on the next shutdown, the service is
shut down correctly. Also this handling is not confused anymore with
what the initial startup is and handled correctly even if the playbook
is executed multiple times. Another advantage is lowerd requirements
on the target machine. Python libraries for docker and docker-compose
need not be installed anymore since ansible is not using the api of
these components anymore.